Update rsi_ble_heartrate.c
In Heart Rate Profile, the values of Body Sensor Location are defined incorrectly. For example, if RSI_BLE_BODY_SENSOR_LOCATION_WRIST is selected, the resulting value of BIT(2) is 4, but the value 4 actually represents Hand. Wrist must be defined as 2 instead of BIT(2). All correct values are as follows:
0 = other
1 = chest
2 = wrist
3 = finger
4 = hand
5 = ear lobe
6 = foot
Verified each value using the nRF Connect Mobile App (Android).